refactor: switch from double to decimal columns
Initial motivation: be able to use PSQL round function (which works on
decimal but not double) when computing nutrient profiles. So using
decimal over double in theory comes with a performance slowdown for
computations, but I imagine this is negligible.

+++ database/migrations/2023_03_29_083314_create_units_table.php
@@ -13,16 +13,16 @@ return new class extends Migration {
         Schema::create('units', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->double('g', 10, 4)->nullable();
-            $table->double('ml', 10, 4)->nullable();
+            $table->decimal('g', 10, 4)->nullable();
+            $table->decimal('ml', 10, 4)->nullable();
             $table->integer('seq_num');
             // ingredient_id is added after ingredients table is created
             // meal_id is added after meals table is created
-            $table->double('custom_unit_amount', 10, 4)->nullable();
-            $table->double('custom_mass_amount', 10, 4)->nullable();
+            $table->decimal('custom_unit_amount', 10, 4)->nullable();
+            $table->decimal('custom_mass_amount', 10, 4)->nullable();
             $table->foreignId('custom_mass_unit_id')->nullable()->references('id')->on('units');
             $table->index('custom_mass_unit_id');
-            $table->double('custom_grams', 10, 4)->nullable();
+            $table->decimal('custom_grams', 10, 4)->nullable();
         });
     }
